AdminOrderDetails Commit

+++ app/controllers/public/orders_controller.rb
@@ -33,17 +33,20 @@ class Public::OrdersController < ApplicationController
   def create
     @cart_items = current_customer.cart_items
     @order = Order.new(order_params)
-    @order.save
-    current_customer.cart_items.each do |cart_item|
-      @order_detail = OrderDetail.new
-      @order_detail.order_id =  @order.id
-      @order_detail.item_id = cart_item.item.id
-      @order_detail.price = cart_item.item.price
-      @order_detail.amount = cart_item.amount
-      @order_detail.save
+    if @order.save
+      current_customer.cart_items.each do |cart_item|
+        @order_detail = OrderDetail.new
+        @order_detail.order_id =  @order.id
+        @order_detail.item_id = cart_item.item.id
+        @order_detail.price = cart_item.item.price
+        @order_detail.amount = cart_item.amount
+        @order_detail.save
+      end
+      current_customer.cart_items.destroy_all
+      redirect_to orders_complete_path
+    else
+      render :confirm
     end
-    current_customer.cart_items.destroy_all
-    redirect_to orders_complete_path
   end
 
   #ありがとう画面
